YARN-10965. Fix review feedbacks

+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-resourcemanager/src/main/java/org/apache/hadoop/yarn/server/resourcemanager/scheduler/capacity/ResourceVector.java
@@ -28,11 +28,10 @@ import java.util.Map;
 import java.util.Set;
 
 /**
- * Represents a simple resource floating point value storage
- * grouped by resource names.
+ * Represents a simple resource floating point value grouped by resource names.
  */
-public class ResourceVector implements Iterable<Map.Entry<String, Float>> {
-  private final Map<String, Float> resourcesByName = new HashMap<>();
+public class ResourceVector implements Iterable<Map.Entry<String, Double>> {
+  private final Map<String, Double> resourcesByName = new HashMap<>();
 
   /**
    * Creates a new {@code ResourceVector} with all pre-defined resources set to
@@ -54,7 +53,7 @@ public class ResourceVector implements Iterable<Map.Entry<String, Float>> {
    * @param value the value to set all resources to
    * @return uniform resource vector
    */
-  public static ResourceVector of(float value) {
+  public static ResourceVector of(double value) {
     ResourceVector emptyResourceVector = new ResourceVector();
     for (ResourceInformation resource : ResourceUtils.getResourceTypesArray()) {
       emptyResourceVector.setValue(resource.getName(), value);
@@ -84,7 +83,7 @@ public class ResourceVector implements Iterable<Map.Entry<String, Float>> {
    * @param otherResourceVector rhs resource vector of the subtraction
    */
   public void decrement(ResourceVector otherResourceVector) {
-    for (Map.Entry<String, Float> resource : otherResourceVector) {
+    for (Map.Entry<String, Double> resource : otherResourceVector) {
       setValue(resource.getKey(), getValue(resource.getKey()) - resource.getValue());
     }
   }
@@ -94,7 +93,7 @@ public class ResourceVector implements Iterable<Map.Entry<String, Float>> {
    * @param resourceName name of the resource
    * @param value value to be subtracted from the resource's current value
    */
-  public void decrement(String resourceName, float value) {
+  public void decrement(String resourceName, double value) {
     setValue(resourceName, getValue(resourceName) - value);
   }
 
@@ -103,15 +102,15 @@ public class ResourceVector implements Iterable<Map.Entry<String, Float>> {
    * @param resourceName name of the resource
    * @param value value to be added to the resource's current value
    */
-  public void increment(String resourceName, float value) {
+  public void increment(String resourceName, double value) {
     setValue(resourceName, getValue(resourceName) + value);
   }
 
-  public float getValue(String resourceName) {
+  public double getValue(String resourceName) {
     return resourcesByName.get(resourceName);
   }
 
-  public void setValue(String resourceName, float value) {
+  public void setValue(String resourceName, double value) {
     resourcesByName.put(resourceName, value);
   }
 
@@ -124,7 +123,7 @@ public class ResourceVector implements Iterable<Map.Entry<String, Float>> {
   }
 
   @Override
-  public Iterator<Map.Entry<String, Float>> iterator() {
+  public Iterator<Map.Entry<String, Double>> iterator() {
     return resourcesByName.entrySet().iterator();
   }
